+++ lib/canql/nodes/dates.rb
@@ -3,7 +3,7 @@
 require 'chronic'
 require 'ndr_support/daterange'
 
-module Canql #:nodoc: all
+module Canql # :nodoc: all
   module Nodes
     module FuzzyDateNode
       delegate :to_daterange, to: :date
@@ -25,12 +25,12 @@ module Canql #:nodoc: all
     module YearQuarterNode
       def to_daterange
         quarter = text_value[0..1]
-        year = text_value[3..-1]
+        year = text_value[3..]
         quarters = {
-          'q1': "01-04-#{year}",
-          'q2': "01-07-#{year}",
-          'q3': "01-10-#{year}",
-          'q4': "01-01-#{year.to_i + 1}"
+          q1: "01-04-#{year}",
+          q2: "01-07-#{year}",
+          q3: "01-10-#{year}",
+          q4: "01-01-#{year.to_i + 1}"
         }
         Daterange.new(quarters[quarter.to_sym].to_date,
                       quarters[quarter.to_sym].to_date + 3.months - 1.day)