+++ lib/canql/nodes/patient_events.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-module Canql #:nodoc: all
+module Canql # :nodoc: all
   module Nodes
     module PatientEvents
       module WithCondition
         def to_events
-          event_name = event_type_name()
+          event_name = event_type_name
           { "#{event_name}_event" => existence_filter }
         end
 
@@ -28,7 +28,7 @@ module Canql #:nodoc: all
         end
 
         def existence_filter
-          { Canql::EQUALS => existance_modifier.text_value.strip != 'no' }
+          { Canql::EQUALS => existence_modifier.text_value.strip != 'no' }
         end
       end
     end