+++ test/nodes/test_result_groups_test.rb
@@ -185,7 +185,7 @@ class TestResultGroupsTest < Minitest::Test
 
     assert parser.valid?
 
-    assert_test_accesptance_count parser, 1
+    assert_test_acceptance_count parser, 1
     assert_test_acceptance_values(
       parser, 0,
       'exists' => { Canql::EQUALS => false },
@@ -200,7 +200,7 @@ class TestResultGroupsTest < Minitest::Test
 
     assert parser.valid?
 
-    assert_test_accesptance_count parser, 1
+    assert_test_acceptance_count parser, 1
     assert_test_acceptance_values(
       parser, 0,
       'exists' => { Canql::EQUALS => false },
@@ -215,7 +215,7 @@ class TestResultGroupsTest < Minitest::Test
 
     assert parser.valid?
 
-    assert_test_accesptance_count parser, 1
+    assert_test_acceptance_count parser, 1
     assert_test_acceptance_values(
       parser, 0,
       'exists' => { Canql::EQUALS => false },
@@ -230,7 +230,7 @@ class TestResultGroupsTest < Minitest::Test
 
     assert parser.valid?
 
-    assert_test_accesptance_count parser, 1
+    assert_test_acceptance_count parser, 1
     assert_test_acceptance_values(
       parser, 0,
       'exists' => { Canql::EQUALS => false },
@@ -260,15 +260,15 @@ class TestResultGroupsTest < Minitest::Test
     assert_dir_block_values(parser, 'test_acceptances', %w[exists acceptance required], index, expected)
   end
 
-  def assert_test_accesptance_count(parser, numder_of_blocks)
-    assert_dir_block_count(parser, 'test_acceptances', numder_of_blocks)
+  def assert_test_acceptance_count(parser, number_of_blocks)
+    assert_dir_block_count(parser, 'test_acceptances', number_of_blocks)
   end
 
   def assert_test_result_group_values(parser, index = 0, expected = {})
     assert_dir_block_values(parser, 'test_result_groups', %w[exists group required], index, expected)
   end
 
-  def assert_test_result_group_count(parser, numder_of_blocks)
-    assert_dir_block_count(parser, 'test_result_groups', numder_of_blocks)
+  def assert_test_result_group_count(parser, number_of_blocks)
+    assert_dir_block_count(parser, 'test_result_groups', number_of_blocks)
   end
 end