+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## [Unreleased]
+## Fixed
+* Code only misspelling of existence
+* Code only misspelling of modifier
+* Code only misspelling of number
+* Code only misspelling of acceptance
 
 ## 6.0.0 / 2024-09-30
 ### Changed